Remove link from top menu items
Removed the href attribute from the top menu items.
On mobile devices there is no hover and hence every
touch leads to a click accessing the ressource behind href.
This prevents that the menu is popping out.

With href removed the menu entry is click/touchable and the
menu expands properly on mobile devices